+++ src/mycrypto/base32.h
@@ -8,13 +8,15 @@ extern "C" {
 #include <stdint.h>
 
 /**
- * @brief Encode a byte array into a printable and null-terminated base32 char array
+ * @brief Encode a byte array into a printable and null-terminated base32 char
+ * array
  * @param input_bytes Pointer to an array of bytes
  * @param input_len Length of input_bytes
- * @returns Pointer to a null-terminated char array or NULL if fails to allocate enough memory,
- * users are reminded to free() the pointer after use.
+ * @returns Pointer to a null-terminated char array or NULL if fails to allocate
+ * enough memory, users are reminded to free() the pointer after use.
  */
-char* encode_bytes_to_base32_string(const uint8_t* input_bytes, const size_t input_len);
+char *encode_bytes_to_base32_string(const uint8_t *input_bytes,
+                                    const size_t input_len);
 
 /**
  * @brief Decode a null-terminated base32-encoded char array into original bytes
@@ -22,10 +24,12 @@ char* encode_bytes_to_base32_string(const uint8_t* input_bytes, const size_t inp
  * @param output_len Pointer to the length of the output byte array. On error:
  * -1 if input_chars is not a valid base32 string.
  * -2 if malloc() failed.
- * @returns Pointer to a byte array or NULL (if error occurs), users are reminded to free() the pointer after use.
- * Note that the byte array returned is NOT guaranteed to be null-terminated.
+ * @returns Pointer to a byte array or NULL (if error occurs), users are
+ * reminded to free() the pointer after use. Note that the byte array returned
+ * is NOT guaranteed to be null-terminated.
  */
-uint8_t* decode_base32_string_to_bytes(const char* input_chars, int64_t* output_len);
+uint8_t *decode_base32_string_to_bytes(const char *input_chars,
+                                       ssize_t *output_len);
 
 #ifdef __cplusplus
 }
